Add UniversalDefaultBoard module
Since classes like EmptyPiece and Move will be creating many instances,
with each needing a reference to the same board, this module will fill
that need.

A class variable is being used instead of a class instance variable
because all of these classes, subclasses, and objects need to have a
reference to the same board.  This way, it just gets set in one place,
and the default value will cascade to everything else that needs it.