Modify EmptyPiece#piece_str
The color formatting was messing up how the text lined up in
BoardDisplay. The text is centered on the square using the String#center
method, but the color format characters count when it comes to computing
the string size.  The font color has been removed from piece_str and
will now need to be handled outside of the class.