Add piece font_color to the row being printed
The font_color was removed from EmptyPiece#piece_str because it was
messing up  the String#center call within the same method.  The center
takes into account all characters in the string, including formatting
characters that do not get directly displayed to the terminal.  So now
the font_color cannot be part of the string being centered and must be
applied separately on the string concatenation line.