Add SingleMovePiece class
King and Knight will now inherit from this class.  Each piece will
calculate all of its possible moves.  Since the algorithm for doing
this would be the same for King and Knight, this parent class was
created to handle that.

Note:  King and Knight each have their own offsets, so they do not move
the same on the board.  The algorithm would calculate the possible moves
by looking at these offsets; the algorithm would be the same.