Add MultipleMovePiece class
This class is to be inherited by pieces that move in a straight line on
a chess board, by at least one space.  Classes that fit this description
are the Queen, Bishop, and Rook.

Each of the chess piece classes will calculate their own possible valid
moves.  Since the algorithm is the same for the Queen, Bishop, and Rook,
this class was created to handle that.  They each have their own
offsets, but that's what the algorithm looks at when generating the
valid move set.